# dist/src/swagger.json
{
	"components": {
...
		"schemas": {
			"SuccessResponse_string_": {
				"properties": {
					"data": {
						"type": "string"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"data",
					"statusCode",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ErrorResponse": {
				"properties": {
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"statusCode",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"JournalPromptCommand": {
				"properties": {
					"id": {
						"type": "string"
					},
					"prompt": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"prompt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"JournalPromptDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"prompt": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"prompt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SuccessResponse_JournalPromptDto_": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/JournalPromptDto"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"data",
					"statusCode",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"JournalEntryCommand": {
				"properties": {
					"id": {
						"type": "string"
					},
					"userId": {
						"type": "string"
					},
					"promptId": {
						"type": "string"
					},
					"entry": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"userId",
					"promptId",
					"entry"
				],
				"type": "object",
				"additionalProperties": false
			},
			"JournalEntryDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"userId": {
						"type": "string"
					},
					"prompt": {
						"type": "string"
					},
					"entry": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"userId",
					"prompt",
					"entry"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SuccessResponse_JournalEntryDto-Array_": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/JournalEntryDto"
						},
						"type": "array"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"data",
					"statusCode",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SuccessResponse_JournalEntryDto_": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/JournalEntryDto"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"data",
					"statusCode",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "athena",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/smartJournal/savePrompt": {
			"post": {
				"operationId": "UpsertPromptAsync",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/SuccessResponse_string_"
										},
										{
											"$ref": "#/components/schemas/ErrorResponse"
										}
									]
								}
							}
						}
					},
					"400": {
						"description": "An error occurred",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Prompt not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "\nCreates and saves journal prompts to the database",
				"tags": [
					"SmartJournalController"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/JournalPromptCommand"
							}
						}
					}
				}
			}
		},
		"/smartJournal/getPrompt/{id}": {
			"get": {
				"operationId": "GetPromptAsync",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/SuccessResponse_JournalPromptDto_"
										},
										{
											"$ref": "#/components/schemas/ErrorResponse"
										}
									]
								}
							}
						}
					},
					"400": {
						"description": "An error occurred",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Prompt Not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "\nRetrieves a specific journal prompt with its appropriate id",
				"tags": [
					"SmartJournalController"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/smartJournal/deletePrompt/{id}": {
			"delete": {
				"operationId": "DeletePromptAsync",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ErrorResponse"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"400": {
						"description": "An error occurred",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Entry Not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Deletes a specific journal prompt with its appropriate id",
				"tags": [
					"SmartJournalController"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/smartJournal/saveEntry": {
			"post": {
				"operationId": "UpsertAsync",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/SuccessResponse_string_"
										},
										{
											"$ref": "#/components/schemas/ErrorResponse"
										}
									]
								}
							}
						}
					},
					"400": {
						"description": "An error occurred",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Entry Not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "\nCreates and saves journal entries to the database",
				"tags": [
					"SmartJournalController"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/JournalEntryCommand"
							}
						}
					}
				}
			}
		},
		"/smartJournal/getEntries/{userId}": {
			"get": {
				"operationId": "GetEntriesAsync",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/SuccessResponse_JournalEntryDto-Array_"
										},
										{
											"$ref": "#/components/schemas/ErrorResponse"
										}
									]
								}
							}
						}
					},
					"400": {
						"description": "An error occurred",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "\nRetrieves journal entries of a specific user",
				"tags": [
					"SmartJournalController"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/smartJournal/getEntries/{userId}/getEntry/{id}": {
			"get": {
				"operationId": "GetEntryAsync",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/SuccessResponse_JournalEntryDto_"
										},
										{
											"$ref": "#/components/schemas/ErrorResponse"
										}
									]
								}
							}
						}
					},
					"400": {
						"description": "An error occurred",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Entry Not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "\nRetrieves a specific journal entry for a user",
				"tags": [
					"SmartJournalController"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/smartJournal/delete/{userId}/deleteEntry/{id}": {
			"delete": {
				"operationId": "DeleteEntryAsync",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ErrorResponse"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"204": {
						"description": "Journal entry deleted"
					},
					"400": {
						"description": "An error occurred",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Entry Not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "\nDeletes a specific journal entry of a user",
				"tags": [
					"SmartJournalController"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}